+++ src/chains/index.js
@@ -1,0 +1,7 @@
+const ethereum = require('@getsafle/vault-eth-controller');
+const bsc = require('@getsafle/bsc-wallet-controller');
+
+module.exports = {
+    ethereum,
+    bsc,
+}